Fix edit event page, fix file service

+++ services/fileService.js
@@ -13,17 +13,21 @@ exports.saveImage = async function (baseImage, oldName, newName, imagePath) {
 	if (oldName === newName || !oldName) {
 		return writeImage(baseImage, imagePath, newName, extension);
 	} else {
-		var deletePromise = await deleteImage(imagePath, oldName);
+		var deletePromise = await exports.deleteImage(imagePath, oldName);
     var writePromise = writeImage(baseImage, imagePath, newName, extension);
     return Promise.all([deletePromise, writePromise]);
 	}
 };
 
-exports.deleteImage = async function (name, imagePath) {
+exports.deleteImage = async function (imagePath, name) {
   var directory = path.join("public/images", imagePath);
   var files = await fs.promises.readdir(directory);
   var image = files.find(file => file.split('.')[0] === name);
   return new Promise((resolve, reject) => {
+    if (!image) {
+      resolve();
+      return;
+    }
     fs.unlink(path.join(directory, image), err => {
       if (err) reject(err);
       resolve();
@@ -35,7 +39,7 @@ function writeImage(baseImage, imagePath, name, extension) {
 	const fileType = baseImage.substring("data:".length,baseImage.indexOf("/"));
 	const regex = new RegExp(`^data:${fileType}\/${extension};base64,`, 'gi');
 	const base64Data = baseImage.replace(regex, "");
-	const imageData = new Buffer.from(base64Data, 'base64');
+  const imageData = new Buffer.from(base64Data, 'base64');
 	const fullPath = path.join('./public/images', imagePath, name) + '.' + extension.replace("jpeg", "jpg");
 
 	return new Promise((resolve, reject) => {
